+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/plugin/RestSqlAction.java
@@ -16,6 +16,7 @@
 package com.amazon.opendistroforelasticsearch.sql.plugin;
 
 import com.alibaba.druid.sql.parser.ParserException;
+import com.amazon.opendistroforelasticsearch.sql.antlr.OpenDistroSqlAnalyzer;
 import com.amazon.opendistroforelasticsearch.sql.antlr.SqlAnalysisException;
 import com.amazon.opendistroforelasticsearch.sql.esdomain.LocalClusterState;
 import com.amazon.opendistroforelasticsearch.sql.exception.SQLFeatureDisabledException;
@@ -52,6 +53,7 @@ import java.util.Set;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
 
+import static com.amazon.opendistroforelasticsearch.sql.plugin.SqlSettings.QUERY_ANALYSIS_ENABLED;
 import static com.amazon.opendistroforelasticsearch.sql.plugin.SqlSettings.SQL_ENABLED;
 import static org.elasticsearch.rest.RestStatus.BAD_REQUEST;
 import static org.elasticsearch.rest.RestStatus.OK;
@@ -132,6 +134,8 @@ public class RestSqlAction extends BaseRestHandler {
     private static QueryAction explainRequest(final NodeClient client, final SqlRequest sqlRequest)
             throws SQLFeatureNotSupportedException, SqlParseException {
 
+        performAnalysisIfEnabled(sqlRequest.getSql());
+
         final QueryAction queryAction = new SearchDao(client).explain(sqlRequest.getSql());
         queryAction.setSqlRequest(sqlRequest);
         return queryAction;
@@ -184,5 +188,24 @@ public class RestSqlAction extends BaseRestHandler {
     private boolean isSQLFeatureEnabled() {
         boolean isSqlEnabled = LocalClusterState.state().getSettingValue(SQL_ENABLED);
         return allowExplicitIndex && isSqlEnabled;
+    }
+
+    private static void performAnalysisIfEnabled(String sql) {
+        // Perform analysis for SELECT only for now because of extra code changes required for SHOW/DESCRIBE.
+        if (!isSelectStatement(sql)) {
+            return;
+        }
+
+        LocalClusterState clusterState = LocalClusterState.state();
+        if (clusterState.getSettingValue(QUERY_ANALYSIS_ENABLED)) {
+            OpenDistroSqlAnalyzer analyzer = new OpenDistroSqlAnalyzer(sql);
+            analyzer.analyze(clusterState);
+        }
+    }
+
+    private static boolean isSelectStatement(String sql) {
+        int endOfFirstWord = sql.indexOf(' ');
+        String firstWord = sql.substring(0, endOfFirstWord > 0 ? endOfFirstWord : sql.length());
+        return "SELECT".equalsIgnoreCase(firstWord);
     }
 }
